changed views structure
