+++ config/settings.py
@@ -9,8 +9,10 @@ https://docs.djangoproject.com/en/3.2/topics/settings/
 For the full list of settings and their values, see
 https://docs.djangoproject.com/en/3.2/ref/settings/
 """
-
+import os
 from pathlib import Path
+from dotenv import load_dotenv
+load_dotenv()
 
 # Build paths inside the config like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
@@ -19,9 +21,10 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # See https://docs.djangoproject.com/en/3.2/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-# TODO: REMOVE TO DOT_ENV file
-SECRET_KEY = 'django-insecure-(*ywn2trkdndh95sh9bcj0%agr!rcy*%mhuu@iy6m)3@)9-no='
-NEWS_API_KEY = ''
+
+
+SECRET_KEY = os.getenv('DJANGO_SECRET_KEY')
+NEWS_API_KEY = os.getenv('NEWS_API_KEY')
 
 
 # SECURITY WARNING: don't run with debug turned on in production!
@@ -93,6 +96,10 @@ AUTH_PASSWORD_VALIDATORS = [
     {
         'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator',
     },
+]
+
+if not DEBUG:
+    AUTH_PASSWORD_VALIDATORS += [
     {
         'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator',
     },
