+++ mainapp/urls.py
@@ -9,7 +9,7 @@ app_name = MainappConfig.name
 urlpatterns = [
     path('', IndexView.as_view(), name='index'),
     path('contacts/', ContactsPageView.as_view(), name='contacts'),
-    path('courses/', CoursesPageView.as_view(), name='courses'),
+    path('courses/', CoursesListView.as_view(), name='courses'),
     path('doc_site/', DocPageView.as_view(), name='doc_site'),
     path('news/', NewsPageView.as_view(), name='news'),
     path('news/<int:page>', NewsPageView.as_view(), name='news'),
