+++ mainapp/views/__init__.py
@@ -1,0 +1,7 @@
+from mainapp.views.auth import *
+from mainapp.views.contacts import *
+from mainapp.views.courses import *
+from mainapp.views.doc_site import *
+from mainapp.views.google_redirect import *
+from mainapp.views.index import *
+from mainapp.views.news import *
