+++ mainapp/views/auth/__init__.py
@@ -1,0 +1,1 @@
+from mainapp.views.auth.login import *
