+++ mainapp/views/auth/login.py
@@ -1,0 +1,7 @@
+__all__ = ['LoginPageView']
+
+from django.views.generic import TemplateView
+
+
+class LoginPageView(TemplateView):
+    template_name = 'mainapp/login.html'
