+++ mainapp/views/contacts/__init__.py
@@ -1,0 +1,1 @@
+from mainapp.views.contacts.list import *
