+++ mainapp/views/contacts/list.py
@@ -1,0 +1,7 @@
+__all__ = ['ContactsPageView']
+
+from django.views.generic import TemplateView
+
+
+class ContactsPageView(TemplateView):
+    template_name = 'mainapp/contacts.html'
