+++ mainapp/views/courses/__init__.py
@@ -1,0 +1,2 @@
+from mainapp.views.courses.list import *
+from mainapp.views.courses.detail import *
