+++ mainapp/views/courses/list.py
@@ -1,0 +1,7 @@
+__all__ = ['CoursesListView']
+
+from django.views.generic import TemplateView
+
+
+class CoursesListView(TemplateView):
+    template_name = 'mainapp/courses_list.html'
