+++ mainapp/views/doc_site/__init__.py
@@ -1,0 +1,1 @@
+from mainapp.views.doc_site.detail import *
