+++ mainapp/views/doc_site/detail.py
@@ -1,0 +1,7 @@
+__all__ = ['DocPageView']
+
+from django.views.generic import TemplateView
+
+
+class DocPageView(TemplateView):
+    template_name = 'mainapp/doc_site.html'
