+++ mainapp/views/google_redirect/__init__.py
@@ -1,0 +1,1 @@
+from mainapp.views.google_redirect.detail import *
