+++ mainapp/views/index/__init__.py
@@ -1,0 +1,1 @@
+from mainapp.views.index.main import *
