+++ mainapp/views/index/main.py
@@ -1,0 +1,12 @@
+__all__ = ['IndexView']
+
+from django.views.generic import TemplateView
+
+
+class IndexView(TemplateView):
+    template_name = 'mainapp/index.html'
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context.update({'special_data': "don't tell anyone!"})
+        return context
