+++ mainapp/views/news/__init__.py
@@ -1,0 +1,1 @@
+from mainapp.views.news.list import *
