+++ requirements.txt
@@ -1,2 +1,3 @@
 Django<3.3
 requests
+python-dotenv
